Refactor js files
